refactor: emphasize success/failure in return type
Among all different statuses, the success/failure state is the most
important. It will be checked in the service or controller layer. We
change the return type so that the success/failure check doesn't involve
hardcoded strings.

An alternative would be implementing the result type using a common base
class, and handling the success/failure check behind a method.

We chose the current method because it skips a layer of abstraction that
is not needed now, and the current method is compatible with the
alternative in JS using [property
getters](https://javascript.info/property-accessors).