fix(jest): broken vscode-jest due to change in env setup method
Previously env vars are provided using env files.
This was changed when we were making a dockerfile to directly setting
env vars.

However this change broke the vscode-jest extension, so we are
rolling back the command to run tests only for testing.

The differences in env setup in jest tests should not be a problem
because the tests are testing for logic and integration, and they use
mocked dependencies anyway so the setup isn't expected to be realistic.